server-stream : pimpl
